test(e2e): shared fixtures, configurable baseURL, commit Cargo.lock
- Extract login() helper to e2e/fixtures.ts using test.extend()
- Remove duplicated login() from 5 test files in tests/e2e/
- Replace hardcoded baseURL with PANOPTIKON_URL env var
- Remove Cargo.lock from .gitignore for reproducible builds
